requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
comet_ml==3.56.0
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
datasets==4.6.1
decorator==5.2.1
dill==0.4.0
dulwich==0.25.2
einops==0.8.2
everett==3.1.0
Farama-Notifications==0.0.4
filelock==3.25.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
gcsfs==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-auth==2.48.0
google-auth-oauthlib==1.3.0
google-cloud-core==2.5.0
google-cloud-storage==3.9.0
google-cloud-storage-control==1.10.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
gymnasium==1.2.3
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
minigrid==3.0.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
numpy==2.4.2
oauthlib==3.3.1
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079150/work
pandas==3.0.1
pillow==12.1.1
platformdirs==4.9.2
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.3.2
python-box==6.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
semantic-version==2.10.0
sentry-sdk==2.54.0
setuptools==80.10.2
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wandb==0.25.0
wheel==0.46.3
wrapt==2.1.1
wurlitzer==3.1.1
xxhash==3.6.0
yarl==1.23.0
